+++ clevercsv/__init__.py
@@ -10,5 +10,5 @@ from .dict_read_write import DictReader, DictWriter
 from .exceptions import Error
 from .parser import field_size_limit
 from .read import reader
-from .wrappers import csv2df, read_as_dicts, read_as_lol
+from .wrappers import csv2df, read_as_dicts, read_csv
 from .write import writer