+++ clevercsv/console/commands/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .code import CodeCommand
+from .detect import DetectCommand
+from .standardize import StandardizeCommand
+from .view import ViewCommand